+++ src/components/visualisation/processingTitle.tsx
@@ -7,10 +7,15 @@ import {
   HStack,
   AccordionIcon,
   Stack,
+  useDisclosure,
+  IconButton,
+  Tooltip,
 } from "@chakra-ui/react";
+import { FaSlidersH } from "react-icons/fa";
 import { components } from "schema/main";
 import { recipeTagMap } from "utils/config/parse";
 import { parseDate } from "utils/generic";
+import { JobParamsDrawer } from "./jobParams";
 
 export interface TwoLineTitleProps {
   title: string;
@@ -49,35 +54,49 @@ const RecipeTag = ({ recipe }: { recipe: string }) => {
   return <Tag colorScheme='blue'>{recipeTagMap[recipe]}</Tag>;
 };
 
-const ProcessingTitle = ({ procJob, autoProc, status }: ProcTitleProps) => (
-  <Stack
-    direction={{ base: "column", md: "row" }}
-    borderTop='solid 4px'
-    borderColor='diamond.200'
-    py={1.5}
-    px={3}
-    w='100%'
-    bg='diamond.100'
-  >
-    <HStack gap={{ base: 0, xl: "5vw" }}>
-      <TwoLineTitle title='Processing Job' value={procJob.processingJobId} />
-      <TwoLineTitle title='AutoProc. Program' value={autoProc.autoProcProgramId} />
-      <TwoLineTitle
-        title='Processing Start'
-        value={parseDate(autoProc.processingStartTime ?? "?")}
-      />
-      <TwoLineTitle title='Processing End' value={parseDate(autoProc.processingEndTime ?? "?")} />
-    </HStack>
-    <Spacer />
-    <HStack>
-      <RecipeTag recipe={procJob.recipe} />
-      <Tag colorScheme={jobStatusColour[status]}>{status}</Tag>
+const ProcessingTitle = ({ procJob, autoProc, status }: ProcTitleProps) => {
+  const { isOpen, onOpen, onClose } = useDisclosure();
+
+  return (
+    <Stack
+      direction={{ base: "column", md: "row" }}
+      borderTop='solid 4px'
+      borderColor='diamond.200'
+      py={1.5}
+      px={3}
+      w='100%'
+      bg='diamond.100'
+    >
+      <HStack gap={{ base: 0, xl: "5vw" }}>
+        <TwoLineTitle title='Processing Job' value={procJob.processingJobId} />
+        <TwoLineTitle title='AutoProc. Program' value={autoProc.autoProcProgramId} />
+        <TwoLineTitle
+          title='Processing Start'
+          value={parseDate(autoProc.processingStartTime ?? "?")}
+        />
+        <TwoLineTitle title='Processing End' value={parseDate(autoProc.processingEndTime ?? "?")} />
+      </HStack>
       <Spacer />
-      <AccordionButton aria-label='Show Content' width='auto'>
-        <AccordionIcon />
-      </AccordionButton>
-    </HStack>
-  </Stack>
-);
+      <HStack>
+        <RecipeTag recipe={procJob.recipe} />
+        <Tag colorScheme={jobStatusColour[status]}>{status}</Tag>
+        <Tooltip label='View Job Parameters'>
+          <IconButton
+            icon={<FaSlidersH />}
+            ml='10px'
+            size='sm'
+            onClick={onOpen}
+            aria-label='View Job Parameters'
+          />
+        </Tooltip>
+        <Spacer />
+        <AccordionButton aria-label='Show Content' width='auto'>
+          <AccordionIcon />
+        </AccordionButton>
+      </HStack>
+      {isOpen && <JobParamsDrawer procJobId={procJob.processingJobId} onClose={onClose} />}
+    </Stack>
+  );
+};
 
 export { TwoLineTitle, ProcessingTitle };