+++ src/loaders/spa.ts
@@ -4,7 +4,8 @@ import { DataConfig, SpaCollectionData } from "schema/interfaces";
 import { components } from "schema/main";
 import { client } from "utils/api/client";
 import { buildEndpoint, includePage } from "utils/api/endpoint";
-import { collectionConfig, spaReprocessingFieldConfig } from "utils/config/parse";
+import { parseJobParameters } from "utils/api/response";
+import { collectionConfig } from "utils/config/parse";
 import { parseData } from "utils/generic";
 
 type DataCollection = components["schemas"]["DataCollectionSummary"];
@@ -59,7 +60,9 @@ export interface SpaResponse {
 
 const getSpaData = async (groupId: string) => {
   const response = await client.safeGet(includePage(`dataGroups/${groupId}/dataCollections`, 1, 1));
-  const returnData = {
+
+  // Tick autocalculation by default
+  const returnData: SpaResponse = {
     collection: {
       info: [],
       comments: "",
@@ -67,6 +70,7 @@ const getSpaData = async (groupId: string) => {
       imageDirectory: "?",
     } as SpaCollectionData,
     jobs: null,
+    jobParameters: { items: { performCalculation: true }, allowReprocessing: false },
   };
 
   if (response.status === 200 && response.data.items) {
@@ -100,25 +104,14 @@ const getSpaData = async (groupId: string) => {
         const processingJobId = jobsResponse.data.items[0].ProcessingJob.processingJobId;
         const response = await client.get(`processingJob/${processingJobId}/parameters`);
 
-        const parameters = response.data as Record<string, string>;
-        const legibleParameters: Record<string, string | boolean> = {};
         if (response.status === 200) {
-          for (const [key, value] of Object.entries(parameters)) {
-            const config = spaReprocessingFieldConfig[key];
-            if (config) {
-              let newValue: string | boolean = value;
-
-              if (config.alias === "gainReferenceFile") {
-                newValue = value.split("/").pop()!;
-              } else {
-                newValue = config.isBool ? value === "1" : value;
-              }
-
-              legibleParameters[config.alias] = newValue;
-            } else {
-              legibleParameters[key] = value;
-            }
-          }
+          returnData.jobParameters = {
+            allowReprocessing: response.data.allowReprocessing,
+            items: {
+              ...returnData.jobParameters.items,
+              ...parseJobParameters(response.data.items),
+            },
+          };
         }
 
         // Ignore extraction step
@@ -152,11 +145,7 @@ const getSpaData = async (groupId: string) => {
           return 0;
         });
 
-        return {
-          collection: parsedCollectionData,
-          jobs: jobsList,
-          jobParameters: legibleParameters,
-        };
+        returnData.jobs = jobsList;
       }
     }
   }