+++ src/routes/SPA.tsx
@@ -34,6 +34,7 @@ import { MdRedo } from "react-icons/md";
 import React from "react";
 import { InfoGroup } from "@diamondlightsource/ui-components";
 import { SpaResponse } from "loaders/spa";
+import { CollectionTitle } from "components/visualisation/collectionTitle";
 
 const Statistics = React.lazy(() => import("components/spa/statistics"));
 
@@ -74,8 +75,7 @@ const SpaPage = () => {
       <HStack marginBottom={2}>
         <VStack w='100%'>
           <HStack w='100%'>
-            <Heading>Data Collection</Heading>
-            <Tag colorScheme='orange'>SPA</Tag>
+            <CollectionTitle type='SPA' colorScheme='orange' />
             <Spacer />
             <Tooltip label='Run Reprocessing'>
               <Button isDisabled={!loaderData.jobParameters.allowReprocessing} onClick={onOpen}>