+++ src/utils/api/response.tsx
@@ -5,6 +5,7 @@ import { parseDate } from "utils/generic";
 import { beamlineToMicroscope } from "utils/config/table";
 import { ParsedSessionReponse, SessionResponse } from "schema/interfaces";
 
+import { spaReprocessingFieldConfig } from "utils/config/parse";
 
 const setHistogram = (
   endpoint: string,
@@ -45,5 +46,34 @@ const parseSessionData = (item: SessionResponse): ParsedSessionReponse => {
   return newItem;
 };
 
-export { setHistogram, parseSessionData };
+/**
+ * Parse object, and return 'human' name for each parameter key when possible,
+ * and convert boolean fields to boolean types
+ *
+ * @param jobParams Initial job parameters
+ * @returns Modified job parameters
+ */
+const parseJobParameters = (jobParams: Record<string, string | boolean>) => {
+  const legibleParameters: Record<string, string | boolean> = {};
 
+  for (const [key, value] of Object.entries(jobParams)) {
+    const config = spaReprocessingFieldConfig[key];
+    if (config) {
+      let newValue: string | boolean = value;
+
+      if (config.alias === "gainReferenceFile") {
+        newValue = (value as string).split("/").pop()!;
+      } else {
+        newValue = config.isBool ? value === "1" : value;
+      }
+
+      legibleParameters[config.alias] = newValue;
+    } else {
+      legibleParameters[key] = value;
+    }
+  }
+
+  return legibleParameters;
+};
+
+export { setHistogram, parseJobParameters, parseSessionData };