+++ CHANGELOG.rst
@@ -3,6 +3,44 @@ Changelog
 ==========
 
 ++++++++++
+v1.5.2 (25/01/2024)
+++++++++++
+
+**Changed**
+
+- SPA reprocessing form now performs checks for maximum and minimum diameter when not stopping after CTF estimation
+- Moved output values closer to top in motion correction row
+- "Calculate for Me" is ticked by default in reprocessing dialogue
+
+**Fixed**
+
+- Motion correction row now responds to CTF summary data clicks
+
+++++++++++
+v1.5.1 (17/01/2024)
+++++++++++
+
+**Changed**
+
+- Reprocessing enabled
+
+++++++++++
+v1.5.0 (12/12/2023)
+++++++++++
+
+**Added**
+
+- About page for processing information
+
+++++++++++
+v1.4.0 (12/12/2023)
+++++++++++
+
+**Added**
+
+- Job parameter overview for processing jobs
+
+++++++++++
 v1.3.5 (27/11/2023)
 ++++++++++